+++ next_action/pick_action.py
@@ -21,21 +21,25 @@ def next_actions(tasks: Tasks, arguments: argparse.Namespace) -> Tasks:
     excluded_projects = arguments.excluded_projects
     # First, get the potential next actions by filtering out completed tasks and tasks with a future creation date or
     # future threshold date
-    actionable_tasks = [task for task in tasks if task.is_actionable()]
+    eligible_tasks = filter(Task.is_actionable, tasks)
     # Then, exclude tasks that have an excluded context
-    eligible_tasks = filter(
-        lambda task: not excluded_contexts & task.contexts() if excluded_contexts else True, actionable_tasks)
+    if excluded_contexts:
+        eligible_tasks = filter(lambda task: not excluded_contexts & task.contexts(), eligible_tasks)
     # And, tasks that have an excluded project
-    eligible_tasks = filter(
-        lambda task: not excluded_projects & task.projects() if excluded_projects else True, eligible_tasks)
+    if excluded_projects:
+        eligible_tasks = filter(lambda task: not excluded_projects & task.projects(), eligible_tasks)
     # Then, select the tasks that belong to all given contexts, if any
-    eligible_tasks = filter(lambda task: contexts <= task.contexts() if contexts else True, eligible_tasks)
+    if contexts:
+        eligible_tasks = filter(lambda task: contexts <= task.contexts(), eligible_tasks)
     # Next, select the tasks that belong to at least one of the given projects, if any
-    eligible_tasks = filter(lambda task: projects & task.projects() if projects else True, eligible_tasks)
+    if projects:
+        eligible_tasks = filter(lambda task: projects & task.projects(), eligible_tasks)
     # If the user only wants to see overdue tasks, filter out non-overdue tasks
-    eligible_tasks = filter(lambda task: task.is_overdue() if arguments.overdue else True, eligible_tasks)
+    if arguments.overdue:
+        eligible_tasks = filter(Task.is_overdue, eligible_tasks)
     # If the user only wants to see tasks due before a due date, filter out non-due tasks
-    eligible_tasks = filter(lambda task: task.is_due(arguments.due) if arguments.due else True, eligible_tasks)
+    if arguments.due:
+        eligible_tasks = filter(lambda task: task.is_due(arguments.due), eligible_tasks)
     # If the user specified a minimum priority, filter out tasks with a lower priority or no priority
     eligible_tasks = filter(lambda task: task.priority_at_least(arguments.priority), eligible_tasks)
     # Remove blocked tasks