+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unrelease]
+
+### Added
+
+- Warn user if there are no next actions because they are using contexts or projects not present in the todo.txt file.
+
 ## [1.1.0] - 2018-06-11
 
 ### Added